--- surface/kasey_task_selector/task_selector_action_client.py
@@ -1,131 +1,0 @@
-import rclpy
-from rclpy.action import ActionClient
-from rclpy.node import Node
-from action_msgs.msg import GoalStatus
-
-# import actions here
-from action_interfaces.action import Example
-
-class ExampleActionClient(Node):
-
-    def __init__(self):
-        # creation of a Node with its name as input
-        super().__init__('example_client')
-        # instantiates new action client with inputs of node, action type, action name
-        self._action_client = ActionClient(self, Example, 'cool_example')
-    
-    # Waits for server to become available, then sends goal, returns a future
-    def send_goal(self, order):
-        self.get_logger().info('Waiting for action server...')
-        self._action_client.wait_for_server()
-
-        goal_msg = Example.Goal()
-        goal_msg.order = order
-
-        self.get_logger().info('Sending goal request...')
-
-        self._send_goal_future = self._action_client.send_goal_async(goal_msg, feedback_callback=self.feedback_callback)
-
-        self._send_goal_future.add_done_callback(self.goal_response_callback)
-
-    # Checks if goal was accepted
-    def goal_response_callback(self, future):
-        goal_handle = future.result()
-        if not goal_handle.accepted:
-            self.get_logger().info('Goal rejected')
-            return
-
-        self.get_logger().info('Goal accepted')
-
-        self._get_result_future = goal_handle.get_result_async()
-        self._get_result_future.add_done_callback(self.get_result_callback)
-
-    # Logs result, exits
-    def get_result_callback(self, future):
-        result = future.result().result
-        status = future.result().status
-        if status == GoalStatus.STATUS_SUCCEEDED:
-            self.get_logger().info('Goal succeeded! Result: {0}'.format(result.sequence))
-        else:
-            self.get_logger().info('Goal failed with status: {0}'.format(status))
-        rclpy.shutdown()
-    
-    # Logs feedback from action server
-    def feedback_callback(self, feedback_msg):
-        feedback = feedback_msg.feedback
-        self.get_logger().info('Received feedback: {0}'.format(feedback.partial_sequence))
-
-    def cancel_goal(self):
-        self.get_logger().info('Canceling goal')
-        # Cancel the goal
-        future = self.goal_handle.cancel_goal_async()
-        future.add_done_callback(self.cancel_done)
-    
-    # Logs if goal was canceled
-    def cancel_done(self, future):
-        cancel_response = future.result()
-        if len(cancel_response.goals_canceling) > 0:
-            self.get_logger().info('Goal successfully canceled')
-        else:
-            self.get_logger().info('Goal failed to cancel')
-        rclpy.shutdown()
-
-"""
-class ____ActionClient(Node):
-
-    def __init__(self):
-        super().__init__('name_action_client')
-        # construct action clients using
-        self._action_client = ActionClient(self, action type, action name)
-
-    def send_goal(self, order):
-        goal_msg = Search.Goal()
-        goal_msg.order = order
-
-        self._action_client.wait_for_server()
-
-        self._send_goal_future = self._action_client.send_goal_async(goal_msg, feedback_callback=self.feedback_callback)
-
-        self._send_goal_future.add_done_callback(self.goal_response_callback)
-
-    def goal_response_callback(self, future):
-        goal_handle = future.result()
-        if not goal_handle.accepted:
-            self.get_logger().info('Goal rejected')
-            return
-
-        self.get_logger().info('Goal accepted')
-
-        self._get_result_future = goal_handle.get_result_async()
-        self._get_result_future.add_done_callback(self.get_result_callback)
-
-    def get_result_callback(self, future):
-        result = future.result().result
-        self.get_logger().info('Result: {0}'.format(result.sequence))
-        rclpy.shutdown()
-    
-    def feedback_callback(self, feedback_msg):
-        feedback = feedback_msg.feedback
-        self.get_logger().info('Received feedback: {0}'.format(feedback.partial_sequence))
-"""
-
-def main(args=None):
-    rclpy.init(args=args)
-
-    action_client_search = ExampleActionClient()
-
-    action_client_search.send_goal(False)
-
-    # create other action clients
-
-    # Get the global executor
-    # rclpy.spin() uses this under the hood & rclpy.shutdown() will shut this executor down
-    global_executor = rclpy.get_global_executor()
-    global_executor.add_node(action_client_search)
-    global_executor.spin()
-
-    # different rclpy spins
-
-
-if __name__ == '__main__':
-    main()
